rename directory (robot_parts)
