+++ Lesson10/src/main/java/by/tms/model/robot_parts/samsung/SamsungHand.java
@@ -1,7 +1,7 @@
-package by.tms.model.robotparts.samsung;
+package by.tms.model.robot_parts.samsung;
 
 import by.tms.interfaces.IHand;
-import by.tms.model.robotparts.RobotParts;
+import by.tms.model.robot_parts.RobotParts;
 import lombok.Getter;
 import lombok.ToString;
 import lombok.experimental.SuperBuilder;
