+++ Lesson10/src/main/java/by/tms/model/robot_parts/samsung/SamsungLeg.java
@@ -1,7 +1,7 @@
-package by.tms.model.robotparts.sony;
+package by.tms.model.robot_parts.samsung;
 
 import by.tms.interfaces.ILeg;
-import by.tms.model.robotparts.RobotParts;
+import by.tms.model.robot_parts.RobotParts;
 import lombok.Getter;
 import lombok.ToString;
 import lombok.experimental.SuperBuilder;
@@ -9,9 +9,9 @@ import lombok.experimental.SuperBuilder;
 @Getter
 @ToString
 @SuperBuilder
-public class SonyLeg extends RobotParts implements ILeg {
+public class SamsungLeg extends RobotParts implements ILeg {
     @Override
     public String go() {
-        return "the leg moves cinematically";
+        return "the leg moves confidently";
     }
 }
